# perf: improve performance of baking