--- a/baker/GrapherBakingUtils.ts
+++ b/baker/GrapherBakingUtils.ts
@@ -81,13 +81,11 @@ export const bakeGrapherUrls = async (urls: string[]) => {
     }
 
     if (toBake.length > 0) {
-        for (const grapherUrls of lodash.chunk(toBake, 5)) {
-            await bakeGraphersToSvgs(
-                grapherUrls,
-                `${BAKED_SITE_DIR}/exports`,
-                OPTIMIZE_SVG_EXPORTS
-            )
-        }
+        await bakeGraphersToSvgs(
+            toBake,
+            `${BAKED_SITE_DIR}/exports`,
+            OPTIMIZE_SVG_EXPORTS
+        )
     }
 }