--- a/baker/siteRenderers.tsx
+++ b/baker/siteRenderers.tsx
@@ -621,7 +621,7 @@ export const renderExplorerPage = async (
         config.id = row.id // Ensure each grapher has an id
         config.adminBaseUrl = ADMIN_BASE_URL
         config.bakedGrapherURL = BAKED_GRAPHER_URL
-        return new Grapher(config).toObject()
+        return config
     }
     const grapherConfigs = grapherConfigRows.map(parseGrapherConfigFromRow)
     const partialGrapherConfigs = partialGrapherConfigRows