Multiple optimisations in here:
- Explorers don't instantiate a `Grapher` object for each `grapherId` contained in it, which then requires downloading the data and metadata file for each of them [150s -> 30s]
- Rudimentarily parallelise `bakeEmbeds` some more by using `p-map` in a single thread [575s -> 412s on my local machine; mostly relevant for a first-time build]